Replace Complex64 with f64

+++ crates/majorana/src/propagator.rs
@@ -1,6 +1,5 @@
 use pyo3::prelude::*;
 
-use num_complex::Complex64;
 use propaq_core::propagator::{AbstractPropagator, PropagationResult};
 use propaq_core::truncators::{reject_surrogate_only, resolve_truncation, FlushSchedule};
 
@@ -21,7 +20,7 @@ use crate::termsum::MajoranaTermSum;
 ///     logger: Optional Logger for verbose JSON Lines event logging.
 #[pyclass(module = "propaq._rust_core")]
 pub struct MajoranaPropagator {
-    inner: AbstractPropagator<MajoranaMonomial, Complex64>,
+    inner: AbstractPropagator<MajoranaMonomial, f64>,
 }
 
 #[pymethods]
